Remove CSS and add jQuery to set Featured Image height to 0
Hiding a Featured Image with CSS removes the image. But this causes the
text in the article to shift up too high. The theme needs to think the
Featured Image is still there so the bottom margin of the Featured
Image still
acts as the top "padding" of the article (this "padding" is actually
the bottom margin of the Featured Image.) Setting the height of the
Featured Image to 0 with jQuery
accomplishes this.